\name{BSWeather}
\alias{BSWeather}
\title{Weather in Basle}
\description{The file contains monthly homogenized values measured at the station located in Basel-Binningen.
}

\encoding{UTF-8}

\usage{data("BSWeather")
}

\format{
A monthly time series from 1864(1) to 2019(6).
 \describe{
    \item{temp}{Average temperature,}
    \item{precip}{Total precipitation.}
 }
}

\details{
For more details please contact the Federal Office of Meteorology and Climatology MeteoSwiss of Switzerland \url{www.meteosuisse.admin.ch}. 


Station:                  Basel / Binningen
Altitude [m asl]:         316 m
Coordinates:              47° 32.5' N / 7° 35.0' E

Reference date of homogenization:
         Temperature:     01.12.2008
         Precipitation:   01.12.2007

Last inhomogeneity provisionally corrected:
         Temperature:     no
         Precipitation:   no

Units:
         Temperature:     °C
         Precipitation:   mm

Missing values:           NA
}


\source{
Federal Office of Meteorology and Climatology MeteoSwiss:

\url{http://www.meteoswiss.admin.ch/home/climate/past/homogenous-monthly-data.html}
}

\references{

}

\examples{

}

\keyword{datasets}
